+++ src/main/groovy/com/netflix/gradle/plugins/deb/validation/DebVersionAttributeValidator.groovy
@@ -1,8 +1,10 @@
 package com.netflix.gradle.plugins.deb.validation
 
-class DebVersionValidator implements VersionValidator {
+import com.netflix.gradle.plugins.packaging.validation.SystemPackagingAttributeValidator
+
+class DebVersionAttributeValidator implements SystemPackagingAttributeValidator {
     /**
-     * Per <a href="http://manpages.ubuntu.com/manpages/quantal/man5/deb-version.5.html">Debian manpage</a> an valid
+     * Per <a href="http://manpages.ubuntu.com/manpages/quantal/man5/deb-version.5.html">Debian manpage</a> a valid
      * version has to follow these conventions:
      *
      * The upstream-version may contain only alphanumerics ("A-Za-
@@ -23,5 +25,10 @@ class DebVersionValidator implements VersionValidator {
 
     private boolean matchesExpectedCharacters(String version) {
         version ==~ /[A-Za-z0-9.+:~-]+/
+    }
+
+    @Override
+    String getErrorMessage(String attribute) {
+        "Invalid upstream version '$attribute' - a valid version must start with a digit and only contain [A-Za-z0-9.+:~-]"
     }
 }