+++ src/main/groovy/com/netflix/gradle/plugins/deb/validation/VersionValidator.groovy
@@ -1,5 +1,0 @@
-package com.netflix.gradle.plugins.deb.validation
-
-interface VersionValidator {
-    boolean validate(String version)
-}