+++ src/main/groovy/com/netflix/gradle/plugins/packaging/validation/SystemPackagingTaskPropertiesValidator.groovy
@@ -1,0 +1,7 @@
+package com.netflix.gradle.plugins.packaging.validation
+
+import org.gradle.api.Task
+
+interface SystemPackagingTaskPropertiesValidator<T extends Task> {
+    void validate(T task)
+}